[change] Turning codebase into library
