--- Experiments/digit_configurations.json
@@ -1,210 +1,0 @@
-{
-    "digit_config": {
-        "0": {
-            "fontscale": 30,
-            "size": [
-                17,
-                30
-            ],
-            "start_point": [
-                0,
-                0
-            ],
-            "bbox": [
-                0,
-                4,
-                16,
-                26
-            ]
-        },
-        "1": {
-            "fontscale": 30,
-            "size": [
-                11,
-                30
-            ],
-            "start_point": [
-                0,
-                0
-            ],
-            "bbox": [
-                0,
-                4,
-                8,
-                26
-            ]
-        },
-        "2": {
-            "fontscale": 30,
-            "size": [
-                16,
-                30
-            ],
-            "start_point": [
-                0,
-                0
-            ],
-            "bbox": [
-                0,
-                4,
-                15,
-                26
-            ]
-        },
-        "3": {
-            "fontscale": 30,
-            "size": [
-                16,
-                30
-            ],
-            "start_point": [
-                0,
-                0
-            ],
-            "bbox":  [
-                0,
-                4,
-                15,
-                26
-            ]
-        },
-        "4": {
-            "fontscale": 30,
-            "size": [
-                16,
-                30
-            ],
-            "start_point": [
-                0,
-                0
-            ],
-            "bbox":  [
-                0,
-                4,
-                15,
-                26
-            ]
-        },
-        "5": {
-            "fontscale": 30,
-            "size": [
-                16,
-                30
-            ],
-            "start_point": [
-                0,
-                0
-            ],
-            "bbox":  [
-                0,
-                4,
-                15,
-                26
-            ]
-        },
-        "6": {
-            "fontscale": 30,
-            "size": [
-                16,
-                30
-            ],
-            "start_point": [
-                0,
-                0
-            ],
-            "bbox":  [
-                0,
-                4,
-                15,
-                26
-            ]
-        },
-        "7": {
-            "fontscale": 30,
-            "size": [
-                16,
-                30
-            ],
-            "start_point": [
-                0,
-                0
-            ],
-            "bbox":  [
-                0,
-                4,
-                15,
-                26
-            ]
-        },
-        "8": {
-            "fontscale": 30,
-            "size": [
-                17,
-                30
-            ],
-            "start_point": [
-                0,
-                0
-            ],
-            "bbox":  [
-                0,
-                4,
-                14,
-                26
-            ]
-        },
-        "9": {
-            "fontscale": 30,
-            "size": [
-                17,
-                30
-            ],
-            "start_point": [
-                0,
-                0
-            ],
-            "bbox":  [
-                0,
-                4,
-                16,
-                26
-            ]
-        },
-        " ": {
-            "fontscale": 30,
-            "size": [
-                10,
-                30
-            ],
-            "start_point": [
-                0,
-                0
-            ],
-            "bbox": [
-                0,
-                0,
-                0,
-                0
-            ]
-        }
-    },
-    "common_configs": {
-        "font_file_loc": "D:\\ACRA\\ImageGenerationPipeline\\Experiments\\terminal-grotesque.grotesque-regular.ttf",
-        "background_color": [
-            255,
-            255,
-            255
-        ],
-        "digit_color": [
-            0,
-            0,
-            0
-        ],
-        "bbox_color": [
-            255,
-            82,
-            82
-        ],
-        "bbox_width": 1
-    }
-}
